Rydde opp i FHO-arenafiler
